Rename report-related functions and modules
